Update SPOKE shortname and title (#96)
* Update SPOKE shortname and title

* Revise SPOKE-OKN description for clarity and detail

Updated the description of the SPOKE-OKN knowledge graph to provide a more comprehensive overview of its features and data integration.

* Enhance spoke-genelab description with detailed information

Updated the description to provide more detailed information about the spoke-genelab KG, including its integration with NASA’s OSDR/GeneLab and the types of data included.

* Improve formatting of spoke-genelab description

Reformatted the description for better readability and fixed a line break issue.

* Revise SPOKE description for clarity and detail

Updated the description of the SPOKE knowledge graph to provide a more comprehensive overview of its integration and coverage.

* Add spoke-okn.md documentation file

* Update spoke-okn stats and endpoint URLs

* Add spoke-okn.md documentation file